feat: 답변 작성자가 다를 경우 삭제 불가 조건 추가
질문 삭제 전에 다른 작성자가 답변을 했는지 확인. 비교를 위해 User에 id+userId로 정의. id만 정의하면 null일
경우 제대로 판단이 되지 않기 때문에 현재 변하지 않기로 정해져있는 userId를 함께 비교

+++ src/main/java/com/codessquad/qna/question/QuestionService.java
@@ -1,6 +1,8 @@
 package com.codessquad.qna.question;
 
+import com.codessquad.qna.answer.Answer;
 import com.codessquad.qna.answer.AnswerService;
+import com.codessquad.qna.exception.InsufficientAuthenticationException;
 import com.codessquad.qna.user.UserDTO;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -58,6 +60,14 @@ public class QuestionService {
     @Transactional
     public void delete(Long id, UserDTO currentSessionUser) {
         Question question = readVerifiedQuestion(id, currentSessionUser);
+        List<Answer> answer = question.getAnswers();
+
+        boolean differentWriterExists = answer.stream()
+                .anyMatch(question::isWriterDifferentFrom);
+
+        if (differentWriterExists) {
+            throw new InsufficientAuthenticationException("다른 작성자가 작성한 답변이 있으면 삭제할 수 없습니다.");
+        }
 
         answerService.deleteAll(question.getAnswers());
         question.delete();
